Introduce a common base type for SourceOrdinaryMethodSymbolBase and SourceUserDefinedOperatorSymbolBase to prepare for explict implementations of operators.
Shared code is pulled from derived types to the new base. This is just a refactoring, no behavior changes intended.
